Internalize MonitorAware* Classes to Parent #62
This commit moves the classes MonitorAwareRectangle and
MonitorAwarePoint to Rectangle and Point as a static class and renames
them to WithMonitor to allow them to be used as Rectangle.WithMonitor
and Point.WithMonitor.

contributes to
https://github.com/eclipse-platform/eclipse.platform.swt/issues/62 and
https://github.com/eclipse-platform/eclipse.platform.swt/issues/128
